Move setupRouter to routes dir
